project3/backend/03-machine-learning/feature-engineering-service/requirements.txt
# Feature Engineering Service Requirements

# Core Data Processing
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# ClickHouse
clickhouse-connect==0.6.8

# Technical Analysis
# ta-lib removed - not needed (candlestick patterns done manually, indicators in aggregates table)

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Central Hub SDK
httpx==0.24.1
requests==2.31.0

# Async Support
asyncio==3.4.3
aiohttp==3.8.5

# Logging & Monitoring
python-json-logger==2.0.7

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
tqdm==4.66.1  # Progress bars

# Testing dependencies removed for production builds
